pyproject.toml
[project]
name = "evo-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "asyncio>=3.4.3",
    "evo-files>=0.1.0",
    "evo-objects>=0.1.0",
    "evo-sdk-common>=0.1.2",
    "tabulate>=0.9.0",
]
